object-detection-traking-yolov8/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.0
ultralytics
streamlit
pafy
youtube-dl
Pillow
opencv-python
pytube